+++ internal/controller/dashboard_controller.go
@@ -105,7 +105,6 @@ func (r *DashboardReconciler) SetupWithManager(mgr ctrl.Manager) error {
 			MatchLabels: map[string]string{
 				DashboardSelectorLabelName: DashboardSelectorLabelValue,
 			},
-			// TODO add match expressions to filter by the tenant label instead of the organization annotation
 		})
 	if err != nil {
 		return errors.WithStack(err)